pyproject.toml
[tool.poetry]
name = "iah-backend-api"
version = "0.1.0"
description = ""
authors = ["Daniel Lee"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "0.104.1"
uvicorn = "0.24.0.post1"
pydantic = {extras = ["email"], version = "1.10.18"}
python-dotenv = "1.0.0"
sqlmodel = "0.0.14"
sqlalchemy = "2.0.23"
uuid = "1.30"
asyncpg = "0.29.0"
sqlalchemy-utils = "0.41.1"
ujson = "5.8.0"
greenlet = "3.0.1"
logger = "1.4"
boto3 = "1.33.9"
python-multipart = "0.0.6"
pillow = "10.1.0"
alembic = "1.13.0"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
pyjwt = "2.8.0"
bcrypt = "4.1.1"
stripe = "7.7.0"
passlib = "1.7.4"
librosa = "^0.10.2"
psycopg2 = "2.9.9"
langchain = "0.1.10"
langchain-community = "0.0.25"
langchain-core = "0.1.28"
langchain-experimental = "0.0.53"
langchain-openai = "0.0.8"
langchain-pinecone = "0.0.3"
langchain-text-splitters = "0.0.2"
bs4 = "0.0.2"
pinecone-client = "3.1.0"
aiosmtplib = "3.0.1"
itsdangerous = "2.1.2"
jinja2 = "3.1.4"
colorlog = "6.8.2"
pillow-heif = "0.16.0"
python-docx = "1.1.2"
pymupdf = "1.24.5"
openpyxl = "3.1.4"
pytesseract = "0.3.10"
pdf2image = "1.17.0"
python-pptx = "0.6.23"
python-socketio = "5.9.0"
amazon-transcribe = "^0.6.2"
setuptools = "^75.1.0"
scipy = "^1.14.1"
numpy = "^1.26.0"
faiss-cpu = "^1.9.0"
pgvector = "^0.3.6"
psycopg = "^3.2.3"
psycopg2-binary = "^2.9.10"
celery = "^5.4.0"
redis = "^5.2.0"
aiofiles = "^24.1.0"
elevenlabs = "1.9.0"
soundfile = "^0.12.1"
langfuse = "<2.59.0"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
pytest-mock = "^3.14.0"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
